requirements.txt
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-readers-file>=0.1.0
chromadb>=0.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
pypdf>=4.0.0
python-dotenv>=1.0.0
rich>=13.0.0
httpx>=0.27.0
